SY-4532: Consolidate tracked rack/device cleanup into TestCase teardown
Review follow-ups: create_test_rack/create_test_devices move to TestCase with
deletion in its teardown, ConsoleCase.teardown now chains to super so the whole
hierarchy cooperates, SimulatorCase and RackCase drop their duplicate tracking,
the select_from_dropdown recovery path can no longer escape the retry loop, and
its docstring uses reST fields per the Python toolchain rules.
